# use ValidatorStatuses when processing slashings during epoch transition.