```
before:
Benchmark                                 (validatorsCount)   Mode  Cnt  Score   Error  Units
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.511 ± 0.063  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.494 ± 0.039  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.500 ± 0.043  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.442 ± 0.208  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.524 ± 0.032  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.446 ± 0.111  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.498 ± 0.042  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.497 ± 0.070  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.475 ± 0.044  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.521 ± 0.044  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.462 ± 0.066  ops/s

after:
Benchmark                                 (validatorsCount)   Mode  Cnt  Score   Error  Units
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.524 ± 0.075  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.558 ± 0.073  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.550 ± 0.040  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.539 ± 0.056  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.568 ± 0.047  ops/s
EpochTransitionBenchmark.epochTransition             400000  thrpt   10  1.540 ± 0.058  ops/s
```

Primarily reducing the requirement to go through state.validators list, as going through validatorStatuses is more efficient. Small performance improvement with this change.

Signed-off-by: Paul Harris <paul.harris@consensys.net>

<!-- Thanks for sending a pull request! Please check out our contribution guidelines: -->
<!-- https://github.com/ConsenSys/teku/blob/master/CONTRIBUTING.md -->

## PR Description

## Fixed Issue(s)
<!-- Please link to fixed issue(s) here using format: fixes #<issue number> -->
<!-- Example: "fixes #2" -->

## Documentation

- [X] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [X] I thought about adding a changelog entry, and added one if I deemed necessary.